Updated openstack/openstack Project: openstack/ceilometer  bdd196b5c02b4c57633319f4c9f4125b488661f1
Edit docs for docs.opentack.org/developer/

Add missed steps in the installation process for running tests.
Mongodb and libmysqlclient-dev packages need to be installed for
Ubuntu12.04 and mongodb and mariadb-devel.x86_64 (or .i386)
need to be installed on Fedora20.

This information has been added to:
doc/source/contriburing/source.rst.

DocImpact
Closes-Bug: #1368974
Change-Id: I163879498b37866c8679b7f69ba60abaacc2e017