Updated openstack/openstack Project: openstack/glance  675a39a74faad5f7fe4eb94e1dcf7f4359ed2285
Fix for adopt glance.store library in Glance

The store module is removed from glance project and new glance_store
module is created, but the glance project code was not updated
properly for the required changes.

_get_from_store and _get_size methods raises store.NotFound exception
from glance_store but not caught in the glance api.

Updated exception.NotFound to store.NotFound which returns 404 NotFound
response to the user.

Closes-Bug: #1380552
Change-Id: I068cb90f5db62443115157a027fbdefec9bd4a00