Updated openstack/openstack Project: openstack/heat  dc2c3c16abde83a23d475960bf84fa903f96b989
Refactor cinder volume v2 API changes

The handling of name and description for v2 API is duplicated for
handle_create and handle_update methods. This patch defines a new
method which handles the differences in parameter names thereby
avoiding duplication and keeping the changes restricted to this
method to the extent possible.

Change-Id: I891aeea3fe5d7cb22be27e05ab72857af668e4ed