Updated openstack/openstack Project: openstack-dev/devstack  50381a46e8e09c8743bae8d002499aa846fc671f
Set Glance default_store to rbd in [glance_store]

Glance is moving from [DEFAULT] to [glance_store] for this option.
Since lib/glance sets both, let's also set it in both places for now.

Failing to do this causes g-api to fail to start with error:
"Store for scheme swift not found"

Change-Id: I9e33ababf7c51f5c750f90b8b366b9892bb4c8cd