Updated openstack/openstack Project: openstack/python-openstackclient  87f5da02c3c0db6807aaa829ada711b3e7a27ce6
Fix operation on clouds with availability-zones

In a cloud with AZs, you can get multiple entries back from the service
catalog - one for each AZ and then one that is AZ agnostic that's tied
to the region. If the region_name is plumbed all the way through, this
works as intended.

Change-Id: I3b365ea306e8111fc80830672ae8080a5d1dc8e0