Updated openstack/openstack Project: openstack/nova  fb091484ccad56dfbf370c781dbc322b4931841d
Cleanup tenant networks plugin config creation

Cleans up the os-tenant-networks plugin config creation.
Although written in a way that the config options could be
declared in multiple places in Nova this is no longer the case
and its not something we should do anyway.

This is some prep work before porting the plugin to the
v2.1 API

Partially implements blueprint v2-on-v3-api

Change-Id: I3b8b26035654f2482b3ba9561c404b6a437ec8a2