Updated openstack/openstack Project: openstack/nova  e726a9f9cf195a59252bc459b833d1e3213431b9
VMware: add support for default pbm policy

This is part of the blueprint vmware-spbm-support.
If there is no defined storage policy in the flavor then the default
storage policy will be used. This patch set adds in the definition of
the default storage policy and validates it existence when the service
starts.

DocImpact
    - pbm_enabled - enables storage policy based placement
    - pbm_default_policy - the default policy that will be used if there
      is none in the flavor
    - pbm_wsdl_location (optional) - PBM service WSDL file location URL.
      e.g. file:///opt/SDK/spbm/wsdl/pbmService.wsdl.

Co-authored-by: Radoslav Gerganov <rgerganov@vmware.com>

Change-Id: I1fc5a99c153d815ea735dccda9d2958dba52ec49