Updated openstack/openstack Project: openstack-dev/pbr  f7537c549909ff3b75157c12004422b213d7b629
Adds option for excluding files from autodoc trees

The arguments originally being passed into sphinx.apidoc specify '.'
as the path to index. Unfortunately this includes the setup.py module.
This causes Sphinx to complain or break depending on the configuration.

This patch ignores setup.py by default and allows the project to
override it in their setup.cfg.

Change-Id: I7c164d42a096ba1a0daf314a689cc3252ca81c49
Closes-Bug: #1260495