Updated openstack/openstack Project: openstack/tempest  0e1942e8ea16d1648955d3d9c27461457c1d2834
Run telemetry in javelin if resources.yaml says so

In previous code telemetry checks were always run (if there
were servers to check). If the environment did not have ceilometer
services this would fail.

check_telemetry now checks that there is a true value on the
'telemetry' key in resources.yaml. No configuration is required for
the telemetry checks, so only a key with a value is required.

Change-Id: I84e0f5938ec808537043dba29a896f991779b04c