Updated openstack/openstack Project: openstack/nova  b16f7e08c52fdf8b5c4b019c18a9847d867de2c0
VMware: mark virtual machines as 'belonging' to OpenStack

This patch registers an extension in vCenter that indicates that the
VM's that are run from OpenStack will belong to OpenStack.

If an administrator tries to update the VM via vCenter then they will
get the following message:

"Solution OpenStack manages the selected virtual machine. You should
not modify the virtual machine directly. Use the management console of
the solution to make changes. Do you want to proceed?"

The extension will be called 'org.openstack.compute'. This will make use
of a type called 'instance'.

Closes-bug: #1366506

Change-Id: I1046576c448704841ae8e1800b8390e947b0d457