Updated openstack/openstack Project: openstack/neutron  68e165ea7ddb7116172552b4f7b1f822921d941d
Remove an argument that is never used

This code was creating a dict with a gw_exists key.  I was curious to
know who was interested in receiving this information down the line.
However, no one is ever interested in that key.  It took me some time
to follow this through wondering what was going on and found only dead
ends.

Change-Id: I755d9628ab750b950b33f5dcf32ccf2a9b00800e