Updated openstack/openstack Project: openstack/cinder  51b756c87c9e9e6a3c20ef7cf9a656f71ba6db16
HP 3PAR: Don't ignore extra-specs snap_cpg when missing cpg

When snap_cpg is specified in extra-specs, it should be used.
For some reason, it was being ignored when extra-specs did not
also specify a user cpg.

When using a volume-type, the snapCPG should come from (in this order
of preference):
1. extra specs snap_cpg,
2. extra specs cpg,
3. config hp3par_cpg_snap,
4. config hp3par_cpg.

Change-Id: I649e3dc49f7775a8071229fc315f0ee5f23aa833
Closes-Bug: 1368972