Updated openstack/openstack Project: openstack/python-neutronclient  72c1473f111846a026cb5f9eb97d68b435eba194
Add InvalidIpForNetworkClient exception

When posting an 'attach interface' request to Nova with an invalid
ip for defined network, Nova returns an HTTP 500 error. In fact,
Neutron returns 'InvalidInput' error, but Neutron client is not able
to translate this error to a specific exception. So that a general
'BadRequest' exception is thrown. Neutron client and Nova need a more
specific Neutron error type, in order to address and translate the
error in to a proper Nova exception.

Change-Id: Idd964c33476d7559f642287d577f8a6122d7674c
Partial-Bug: 1369871