Updated openstack/openstack Project: openstack/horizon  9d846d19f637db9906acf80fd6ff5b5b83d57710
Fixes poor error message when deleting networks

This patch fixes the poor error message when deleting networks when
the network is being used by an active instance. The old error
message upon network deletion used to show the network UUID, which
was a little inconvenient for the end-user. So, with the help of
this patch, any error in deleting the network will show the
network name instead of network UUID.

Change-Id: I38a8a0260860383eee9d3e3b28cc4bc3ec65b255
Closes-bug: #1279485