Updated openstack/openstack Project: openstack/nova  45c2205ef13ecb32e41dde0b6c744614d445eb8e
Fix wrong exception return in fixed_ips v2 extention

Currently, fixed_ips api raises HTTPNotFound if passing an invalid
ip address in url, I think this is not correct. It's better to return
HTTPBadRequest to end user when the api excepts a FixedIpInvalid
exception.

This patch fixes it and also the test cases.

Change-Id: I7b18b76745f9092456d511ad023a9ed3a24cca84