Updated openstack/openstack Project: openstack/nova  c3b2ee33cc418e2ded0c5cc19b31eb5723a27fad
Add 'zvm' to the list of known hypervisor types

zvm is a hypervisor that used for IBM s390 arch and has
many years history on virtualization.
you can refer to http://www.vm.ibm.com/ for more info on it.

According to danpb's comments on
https://review.openstack.org/#/c/122200/:
FYI the list of hvtype constants is intended to define a global set of
all theoretically supported hypervisors. ie it is fine to add more
constants for hypervisors, even if Nova doesn't have a driver in-tree for it.

even if zvm is not in the nova tree now, we are trying to enhance the
code and other related stuffs such as CI test to make it acceptable to
community later. Also, zvm driver code is not in stackforge now since
we are trying to fit other criterias in order to make community accept it and
put into nova tree.

This patch simply adds zvm to the list of known hypervisor types which are
validated during scheduling given commit a4fd236.

Change-Id: I1f5b5bd2a8f082d2e945560ae82593dfd70f2a22