Updated openstack/openstack Project: openstack/ceilometer  8db7282fa0715b96aa20ad9b47c681502b77b082
Avoid clobbering existing class definition

Using the same variable name for a new class implies that an
existing class is being modified whereas what's happening is that a
subclass is being created of an existing class and having a
different value for an attribute. Such an implication is confusing
for maintainability.

Change-Id: If784949e8b5706e881bc702ff0b30848bca17d5c