Updated openstack/openstack Project: openstack/glance  c4e19b8f3c945f26df9d0344657e6c6f933b2a5e
Reduce extraneous test output

The tests generate a large amount of extraneous logging output.
This due to the manage utility, when tested via test_manage,
setting up logging in it's own way. Monkey patching the log.setup
call to a no-op during tests prevents this.

Also fixed an incorrect usage of assert which was causing a
warning to be generated.

Change-Id: I6b8724142c0812d84e697e280acc0264620f6a2b