Updated openstack/openstack Project: openstack/keystone  fdbad9f530ea4478d96437b021c9b5cc6d338901
Restrict certain APIs to cloud admin in domain-aware policy

Some of the APIs in the domain-aware policy file are currently
allowed by any "admin" user, when they should really be locked
down to the cloud admin.  Without this, users who are a project
admin will be allowed to do things like manage regions, IdPs,
and other objects that they should not be allowed to touch.

Change-Id: Ifca8bc2fffd2d8c1bf02373d1fadd459a77f836c
Closes-bug: #1381809