Updated openstack/openstack Project: openstack/oslo.config  11ecf187dc380ff5b5d235785765a5e61917ab42
Add missing newline to stderr output when argument value is wrong

For example when having an argument with possible choices and the used
choice is not available the stderr output is like the following example.

argument --formatting: Invalid String(choices=['check', 'fix']) value: fixaberendt@zitrone

To have the shell prompt at the next line it necessary to at '\n' at the
end of the stderr output.

Change-Id: I99152e69bb090ea5d7102cc85d60498c3c6213eb