Updated openstack/openstack Project: openstack/nova  2b238e822ce33697e1dd13e9035bf668c51588fc
Optimize get_instance_nw_info and remove ipam

The get network info was making 4 * num_networks + 1 rpc and db
queries. The network info is recalculated quite often, so we use
the new joins to turn this into a single query. In the process we
remove nova_ipam_lib which was initially created as part of a very
old neutron proxy and only added an unnecessary layer of abstraction.

Change-Id: I34ca575fb33b43ab8780357990b73df837da71c2
Closes-Bug: #1365606