Updated openstack/openstack Project: openstack/nova  04272f8b872456de71abdd573aaf518e64f6c692
Handle Forbidden error from network_api.show_port in os-interface:show

We can get a 403 back from the neutronv2 API when calling
network_api.show_port so handle that in the compute API extension.

This also updates the v3 API code to match the v2 API code to put the
original error message for the 404 into the HTTPNotFound error.

Closes-Bug: #1378389

Change-Id: I7afc832f26481dc44f1c500b6121e2060f36c63a