Updated openstack/openstack Project: openstack/cinder  89ba6909e5de3f76428131f12647ee1b86729c90
Fix display name change during backup restore

Currently the display name and description of the target volume
is changed to that in the backup meta-data during backup restore.
This can confuse the end-user, especially in the case of restore
to an existing volume. This patch removes display name and display
description from the list of fields to be copied from the backup
meta-data to update the volume.

Change-Id: I21a0505352ddc8db801f4820d575399a4771b259
Closes-Bug: #1376194