Updated openstack/openstack Project: openstack/horizon  4382c8f4b2ae27065734d30aa194de6dfd3a47a0
Update WSGI app creation to be compatible with Django 1.7

This resolves the following error when trying to run Horizon:

AppRegistryNotReady: The translation infrastructure cannot be
initialized before the apps registry is ready. Check that you don't
make non-lazy gettext calls at import time.

and is backwards-compatible with our currently supported versions (back
to 1.4).

Change-Id: I1e59f2bf6633a34cfe9230aa028f345e5fa70f94
Closes-Bug: #1382023