Updated openstack/openstack Project: openstack-dev/devstack  77490b5eaefdc8797fd0531fc01c92ccb37400f4
XenAPI: Fix vimrc

DevStack installs a default minimal .vimrc file for the stack user.
Unfortunately the `syntax on` config line is not recognised by the vim
installed in DomU. It results in an annoying message being displayed
whenever the user is using vi. To avoid this issue, removing the `syntax
on` line.

Change-Id: I224465cc3cdba3464ea0a9a751f250ecb6ddc9f3