Updated openstack/openstack Project: openstack/nova  ad84585d750d557712414c5462c47cfcee43aa8f
Fix instance_extra backref

This fixes cruft from the original NUMA patches that placed a backref on
the Instance model for numa_topology. This changes it to the generic
'instance_extra'. Since this was not actually used, nothing breaks; the
extra backref will be used in subsequent patches.

Related to blueprint flavor-from-sysmeta-to-blob

Change-Id: Ifa69e4704ea3f60b4aced3504bf9c599e3868616