Updated openstack/openstack Project: openstack/horizon  d471bae2cc82ace7cc8fe23b0acfbf44a5612803
Do not log keystone token

Previously token values are logged as DEBUG level when a new client
object is instantiated. In other project and clients, token values
are now not logged and is output as *REDACTED* instead.
In Horizon these log lines do not have much meaning and
we can simply remove them.

Change-Id: I67617ac6424907574d79ec2a57b513a548e220d2
Closes-Bug: #1380642