Updated openstack/openstack Project: openstack/python-openstackclient  b40fa49809e8a7d2393429ff140f29b51925999e
Move plugin stuff to clientmanager

The OSC plugins work by adding an object as an attribute to a
ClientManager instance.  The initialization and management of
thos plugins belongs in clientmanager.py.

At this point the only part not moved is the API version dict
initialization bcause the timing and connection to the
CommandManager initialization.  It gets refactored anyway when
API discovery becomes operational.

Change-Id: If9cb9a0c45a3a577082a5cdbb793769211f20ebb