Updated openstack/openstack Project: openstack/nova  f9a0735c81250f1029a7f9895adc629e28d81cbc
Return HTTP 400 if use an in-use fixed ip to attach interface

When posting an 'attach interface' request to Nova with an in-used fixed ip,
Nova need to return HTTP 400, instead of HTTP 500, in order to inform the
user to correct the fixed ip in HTTP request. This code change will take
effect with Neutron only.

Change-Id: Ic5fe49ef0966c981756b5c6e1bf755e587a95ac1
Closes-Bug: 1363901