Updated openstack/openstack Project: openstack/tempest-lib  5715fd65142380573c5cb1f0ff4f3996004b3466
Switch to elapsed time in subunit-trace summary

This commit switches the time output to the elapsed time instead of
the total time in the results summary at the end of subunit-trace.
This is accomplished by just measuring the time delta right before
we begin processing the result stream until right after. While this
is not the most accurate method to measure the elapsed time, it
should be sufficient.

Previously just the total time was shown which could be misleading
since it is just the sum of the execute time for the individual test
methods (excluding the test fixtures). When running in parallel this
number would usually be a fair bit larger than the elapsed time of the
test run. By using the elapsed time instead it makes it easier to track
the performance of the run.

Change-Id: I7a1113c7f1333389b1b42249cafc4a45d630cb9f