Updated openstack/openstack Project: openstack/taskflow  e306f53daefbc93f8c414d3f8bcabfa39cf2ef97
Link a few of the classes to implemented features/bugs in python

This adds comments that associate the classes we have for threading
usage to upstream bugs in python where similar features are being
created (and potentially supported upstream).

When we are able to reduce the number of supported python versions
we can/should try to remove our implementations and move to the ones
that may showup in the python standard library instead.

Change-Id: I7b58380aeb57a58fa3b3c424c9f39de30f44f0e9